--- a/emacs/emacs-config.org
+++ b/emacs/emacs-config.org
@@ -1419,6 +1419,26 @@ This is a collection of extra utilities for various packages.
       (:defer 1 #'akirak-files-ensure-abbrev-list)
       (run-with-idle-timer 300 nil #'github-linguist-update-projects)))
 #+end_src
+**** akirak-elec-pair                                        :parentheses:
+:PROPERTIES:
+:CREATED_TIME: [2022-02-27 Sun 14:39]
+:END:
+#+begin_src emacs-lisp
+  (setup akirak-elec-pair
+    (:status opinionated
+      (:global "M-i" nil
+               "M-i c" #'akirak-elec-pair-replace
+               "M-i d" #'akirak-elec-pair-delete
+               "M-i y" #'akirak-elec-pair-wrap-post-yank
+               "M-i #" #'akirak-elec-pair-self-insert
+               "M-i $" #'akirak-elec-pair-self-insert
+               "M-i %" #'akirak-elec-pair-self-insert
+               "M-i {" #'akirak-elec-pair-self-insert
+               "M-i [" #'akirak-elec-pair-self-insert
+               "M-i (" #'akirak-elec-pair-self-insert)))
+#+end_src
+
+I once used [[https://github.com/cute-jumper/embrace.el][embrace.el]], but it had dependency on expand-region I wanted to get rid of, so I implemented a simpler alternative.
 **** akirak-window                                               :windows:
 :PROPERTIES:
 :CREATED_TIME: [2022-02-24 Thu 14:10]
@@ -2083,6 +2103,8 @@ For Emacs Lisp, I use lispy.
                   lisp-interaction-mode
                   ielm-mode
                   eval-expression-minibuffer-setup)
+
+      (:unbind "M-i")
 
       (:with-map lispy-mode-map-lispy
         ;; Prevent conflicts with other custom keybindings