--- a/emacs/emacs-config.org
+++ b/emacs/emacs-config.org
@@ -2407,6 +2407,14 @@ For configuration, see [[https://git.sr.ht/~tomterl/repl-toggle][~tomterl/repl-t
 #+begin_src emacs-lisp
 (setup (:package string-inflection))
 #+end_src
+*** suggest                          :Emacs__Lisp:development:help:ARCHIVE:
+:PROPERTIES:
+:CREATED_TIME: [2022-02-27 Sun 18:10]
+:END:
+
+#+begin_src emacs-lisp
+(setup (:package suggest))
+#+end_src
 *** svg-lib                                              :library:graphics:
 :PROPERTIES:
 :CREATED_TIME: [2022-01-05 Wed 15:30]