--- a/emacs/emacs-config.org
+++ b/emacs/emacs-config.org
@@ -376,22 +376,20 @@ The built-in =electric= package provides the following modes:
     (:with-mode emacs-lisp-mode
       (:huan ("eval-after-load" "with-eval-after-load"))
       (:local-set imenu-generic-expression
-                  ;; Copied from my previous config. Maybe requires some adaptation.
-                  `(("Headings" ";;[;]\\{1,8\\} \\(.*$\\)" 1)
-                    (nil "^\\s-*(\\(cl-def\\(?:generic\\|ine-compiler-macro\\|m\\(?:acro\\|ethod\\)\\|subst\\|un\\)\\|def\\(?:advice\\|generic\\|ine-\\(?:advice\\|compil\\(?:ation-mode\\|er-macro\\)\\|derived-mode\\|g\\(?:\\(?:eneric\\|lobal\\(?:\\(?:ized\\)?-minor\\)\\)-mode\\)\\|inline\\|m\\(?:ethod-combination\\|inor-mode\\|odify-macro\\)\\|s\\(?:etf-expander\\|keleton\\)\\)\\|m\\(?:acro\\|ethod\\)\\|s\\(?:etf\\|ubst\\)\\|un\\*?\\)\\|ert-deftest\\)\\s-+\\(\\(?:\\sw\\|\\s_\\|\\\\.\\)+\\)" 2)
-                    ("Custom vars" "^\\s-*(\\(defcustom\\)\\s-+\\(\\(?:\\sw\\|\\s_\\|\\\\.\\)+\\)" 2)
-                    ("Variables" "^\\s-*(\\(def\\(?:c\\(?:onst\\(?:ant\\)?\\|ustom\\)\\|ine-symbol-macro\\|parameter\\)\\)\\s-+\\(\\(?:\\sw\\|\\s_\\|\\\\.\\)+\\)" 2)
-                    ("Variables" "^\\s-*(defvar\\(?:-local\\)?\\s-+\\(\\(?:\\sw\\|\\s_\\|\\\\.\\)+\\)[[:space:]\n]+[^)]" 1)
-                    ("Types" "^\\s-*(\\(cl-def\\(?:struct\\|type\\)\\|def\\(?:class\\|face\\|group\\|ine-\\(?:condition\\|error\\|widget\\)\\|package\\|struct\\|t\\(?:\\(?:hem\\|yp\\)e\\)\\)\\)\\s-+'?\\(\\(?:\\sw\\|\\s_\\|\\\\.\\)+\\)" 2)
-                    ;; Only in my configuration
-                    ("use-package" ,(rx "(use-package" (+ space)
-                                        (group (+ (any alnum "-"))))
-                     1)
-                    ("org-capture" ,(rx "(org-starter-def-capture" (+ space)
-                                        (+ (not space)) (+ space)
-                                        (syntax string-quote)
-                                        (group (+ (not (syntax string-quote)))))
-                     1))))
+                  (append lisp-imenu-generic-expression
+                          `(("Misc"
+                             ,(rx bol "("
+                                  (or (and "c" (not (any "l")))
+                                      (any "ab" "d-z"))
+                                  (* (any "-" alnum))
+                                  "-"
+                                  "def" (+ (not (any space)))
+                                  (+ (any blank))
+                                  (group (+ (not (any space control "\n()")))))
+                             1)
+                            ("Config"
+                             "^(\\(?:setup\\|use-package\\)[[:blank:]]+\\_<\\([^[:space:]]+\\)"
+                             1)))))
 
     (:auto-insert
      ("-tests?\\.el\\'" . (> ";;; -*- lexical-binding: t -*-\n\n"