--- a/emacs/emacs-config.org
+++ b/emacs/emacs-config.org
@@ -2748,31 +2748,31 @@ If you have [[https://github.com/alphapapa/org-ql/][org-ql]] installed, you can 
 #+end_src
 
 #+RESULTS: tag-statistics
-| [[org-ql-search:tags:packages][packages]]            | 118 |
+| [[org-ql-search:tags:packages][packages]]            | 120 |
 | [[org-ql-search:tags:visual][visual]]              |  11 |
-| [[org-ql-search:tags:UI][UI]]                  |  10 |
 | [[org-ql-search:tags:formatting][formatting]]          |   9 |
 | [[org-ql-search:tags:development][development]]         |   9 |
+| [[org-ql-search:tags:UI][UI]]                  |   8 |
 | [[org-ql-search:tags:windows][windows]]             |   7 |
 | [[org-ql-search:tags:dired][dired]]               |   6 |
 | [[org-ql-search:tags:completion][completion]]          |   6 |
 | [[org-ql-search:tags:Emacs__Lisp][Emacs__Lisp]]         |   5 |
 | [[org-ql-search:tags:keybindings][keybindings]]         |   5 |
+| [[org-ql-search:tags:Emacs][Emacs]]               |   5 |
 | [[org-ql-search:tags:library][library]]             |   5 |
+| [[org-ql-search:tags:editing][editing]]             |   5 |
 | [[org-ql-search:tags:process][process]]             |   4 |
 | [[org-ql-search:tags:major__mode][major__mode]]         |   4 |
-| [[org-ql-search:tags:Emacs][Emacs]]               |   4 |
 | [[org-ql-search:tags:history][history]]             |   4 |
+| [[org-ql-search:tags:VC][VC]]                  |   4 |
 | [[org-ql-search:tags:profiling][profiling]]           |   3 |
+| [[org-ql-search:tags:parentheses][parentheses]]         |   3 |
 | [[org-ql-search:tags:files][files]]               |   3 |
 | [[org-ql-search:tags:minibuffer][minibuffer]]          |   3 |
 | [[org-ql-search:tags:writing][writing]]             |   3 |
-| [[org-ql-search:tags:VC][VC]]                  |   3 |
 | [[org-ql-search:tags:links][links]]               |   3 |
-| [[org-ql-search:tags:editing][editing]]             |   3 |
 | [[org-ql-search:tags:expansion][expansion]]           |   2 |
 | [[org-ql-search:tags:IO][IO]]                  |   2 |
-| [[org-ql-search:tags:parentheses][parentheses]]         |   2 |
 | [[org-ql-search:tags:Prolog][Prolog]]              |   2 |
 | [[org-ql-search:tags:navigation][navigation]]          |   2 |
 | [[org-ql-search:tags:reading][reading]]             |   2 |