--- a/flake.nix
+++ b/flake.nix
@@ -278,7 +278,7 @@
             update-elisp = channels.nixpkgs.writeShellScriptBin "update-elisp" ''
               nix flake lock --update-input melpa --update-input gnu-elpa
               cd emacs/lock
-              bash ./update.bash
+              bash ./update.bash "$@"
             '';
           }
           //