This PR contains a bunch of small improvements on my Emacs config. Most notably, the startup time has significantly reduced (from 1.0 sec to 0.45 sec in benchmark-init and from 1.8 sec to 0.2 sec in esup).